Moved a class to a separate file. Fixed a file name
